requirements.txt
langchain==0.3.14
langchain_community==0.3.14
langchain_core==0.3.30
langchain_groq==0.2.3
pymongo==4.10.1
python-dotenv==1.0.1
streamlit==1.41.1
Werkzeug==3.1.3
tavily-python==0.5.0
pandas==2.2.3
